[mesh. gf] Refactor evaluate as a free function
- implementation details.
  To minimize the concept of a mesh and make implementation of future
meshes easier.
