+++ triqs/mesh/b_zone.hpp
@@ -102,19 +102,12 @@ namespace triqs::mesh {
       return result;
     }
 
-    template <typename F> auto evaluate(F const &f, std::array<double, 3> const &x) const { return multivar_eval(f, get_interpolation_data(x)); }
-
     // -------------- Evaluation of a function on the grid --------------------------
 
     /// Reduce index modulo to the lattice.
     index_t index_modulo(index_t const &r) const { return index_t{_modulo(r[0], 0), _modulo(r[1], 1), _modulo(r[2], 2)}; }
 
     interpol_data_0d_t<index_t> get_interpolation_data(index_t const &x) const { return {index_modulo(x)}; }
-
-    template <typename F> auto evaluate(F const &f, index_t const &x) const {
-      auto id = get_interpolation_data(x);
-      return f[id.idx[0]];
-    }
 
     // ------------------- Comparison -------------------
